+++ src/collective/cart/core/tests/base.py
@@ -1,4 +1,3 @@
-"""Base module for unittesting"""
 from plone.app.testing import FunctionalTesting
 from plone.app.testing import IntegrationTesting
 from plone.app.testing import PLONE_FIXTURE