+++ src/collective/cart/core/tests/test_ArticleAdapter.py
@@ -15,10 +15,10 @@ class TestArticleAdapter(IntegrationTestCase):
         self.portal = self.layer['portal']
 
     def test_subclass(self):
-        from collective.cart.core.adapter.base import BaseAdapter
+        from collective.base.adapter import BaseAdapter
         from collective.cart.core.adapter.article import ArticleAdapter
         self.assertTrue(issubclass(ArticleAdapter, BaseAdapter))
-        from collective.cart.core.interfaces import IBaseAdapter
+        from collective.base.interfaces import IBaseAdapter
         from collective.cart.core.interfaces import IArticleAdapter
         self.assertTrue(issubclass(IArticleAdapter, IBaseAdapter))