+++ src/collective/cart/core/tests/test_ShoppingSite.py
@@ -1,7 +1,13 @@
 from Testing import ZopeTestCase as ztc
+from collective.cart.core.interfaces import IShoppingSite
+from collective.cart.core.interfaces import IShoppingSiteRoot
 from collective.cart.core.tests.base import IntegrationTestCase
 from plone.app.testing import TEST_USER_ID
 from plone.app.testing import setRoles
+from plone.dexterity.utils import createContentInContainer
+from zope.interface import alsoProvides
+from zope.interface import noLongerProvides
+from zope.lifecycleevent import modified
 
 import mock
 
@@ -14,25 +20,21 @@ class TestShoppingSite(IntegrationTestCase):
         setRoles(self.portal, TEST_USER_ID, ['Manager'])
 
     def test_subclass(self):
-        from collective.cart.core.adapter.base import BaseAdapter
+        from collective.base.adapter import BaseAdapter
         from collective.cart.core.adapter.interface import ShoppingSite
         self.assertTrue(issubclass(ShoppingSite, BaseAdapter))
-        from collective.cart.core.interfaces import IBaseAdapter
-        from collective.cart.core.interfaces import IShoppingSite
+        from collective.base.interfaces import IBaseAdapter
         self.assertTrue(issubclass(IShoppingSite, IBaseAdapter))
 
     def test_provides(self):
         from collective.cart.core.adapter.interface import ShoppingSite
-        from collective.cart.core.interfaces import IShoppingSite
         self.assertEqual(getattr(ShoppingSite, 'grokcore.component.directive.provides'), IShoppingSite)
 
     def test_instance(self):
         from collective.cart.core.adapter.interface import ShoppingSite
-        from collective.cart.core.interfaces import IShoppingSite
         self.assertIsInstance(IShoppingSite(self.portal), ShoppingSite)
 
     def test_instance__provides(self):
-        from collective.cart.core.interfaces import IShoppingSite
         self.assertEqual(getattr(IShoppingSite(self.portal), 'grokcore.component.directive.provides'), IShoppingSite)
 
     def create_folder(self, context=None, oid=None):
@@ -45,9 +47,6 @@ class TestShoppingSite(IntegrationTestCase):
         return folder
 
     def test_shop(self):
-        from collective.cart.core.interfaces import IShoppingSite
-        from collective.cart.core.interfaces import IShoppingSiteRoot
-        from zope.interface import alsoProvides
         folder1 = self.create_folder(oid='folder1')
         folder2 = self.create_folder(folder1, 'folder2')
         folder3 = self.create_folder(folder2, 'folder3')
@@ -64,12 +63,6 @@ class TestShoppingSite(IntegrationTestCase):
         self.assertEqual(IShoppingSite(folder3).shop, folder2)
 
     def test_cart_container(self):
-        from collective.cart.core.interfaces import IShoppingSite
-        from collective.cart.core.interfaces import IShoppingSiteRoot
-        from plone.dexterity.utils import createContentInContainer
-        from zope.interface import alsoProvides
-        from zope.interface import noLongerProvides
-        from zope.lifecycleevent import modified
         self.assertIsNone(IShoppingSite(self.portal).cart_container)
 
         folder = self.create_folder()
@@ -87,7 +80,6 @@ class TestShoppingSite(IntegrationTestCase):
         self.assertIsNone(IShoppingSite(folder).cart_container)
 
     def test_cart(self):
-        from collective.cart.core.interfaces import IShoppingSite
         shopping_site = IShoppingSite(self.portal)
         self.assertIsNone(shopping_site.cart)
 
@@ -96,7 +88,6 @@ class TestShoppingSite(IntegrationTestCase):
         self.assertEqual(shopping_site.cart, 'CART')
 
     def test_cart_articles(self):
-        from collective.cart.core.interfaces import IShoppingSite
         shopping_site = IShoppingSite(self.portal)
         self.assertIsNone(shopping_site.cart_articles)
 
@@ -108,7 +99,6 @@ class TestShoppingSite(IntegrationTestCase):
         self.assertEqual(shopping_site.cart_articles, 'ARTICLES')
 
     def test_cart_article_listing(self):
-        from collective.cart.core.interfaces import IShoppingSite
         shopping_site = IShoppingSite(self.portal)
         self.assertEqual(shopping_site.cart_article_listing, [])
 
@@ -117,7 +107,6 @@ class TestShoppingSite(IntegrationTestCase):
         self.assertEqual(shopping_site.cart_article_listing, ['ARTICLE1', 'ARTICLE2'])
 
     def test_get_cart_article(self):
-        from collective.cart.core.interfaces import IShoppingSite
         shopping_site = IShoppingSite(self.portal)
         self.assertIsNone(shopping_site.get_cart_article('1'))
 
@@ -128,7 +117,6 @@ class TestShoppingSite(IntegrationTestCase):
         self.assertEqual(shopping_site.get_cart_article('2'), 'ARTICLE2')
 
     def test_remove_cart_articles(self):
-        from collective.cart.core.interfaces import IShoppingSite
         shopping_site = IShoppingSite(self.portal)
         session = shopping_site.getSessionData(create=True)
         session.set('collective.cart.core', {'articles': {'1': 'ARTICLE1', '2': 'ARTICLE2', '3': 'ARTICLE3'}})
@@ -145,11 +133,6 @@ class TestShoppingSite(IntegrationTestCase):
     # CartContainer related methods
 
     def test_get_cart(self):
-        from collective.cart.core.interfaces import IShoppingSite
-        from collective.cart.core.interfaces import IShoppingSiteRoot
-        from plone.dexterity.utils import createContentInContainer
-        from zope.interface import alsoProvides
-        from zope.lifecycleevent import modified
         self.assertIsNone(IShoppingSite(self.portal).get_cart('1'))
 
         folder = self.create_folder()
@@ -165,23 +148,25 @@ class TestShoppingSite(IntegrationTestCase):
         self.assertIsNone(IShoppingSite(folder).get_cart('2'))
         self.assertEqual(IShoppingSite(folder).get_cart('1'), cart1)
 
+    def create_cart_container(self, context):
+        container = createContentInContainer(
+            context, 'collective.cart.core.CartContainer', id='cart-container', checkConstraints=False)
+        modified(container)
+        return container
+
     @mock.patch('collective.cart.core.adapter.interface.ICartContainerAdapter')
     def test_update_next_cart_id(self, ICartContainerAdapter):
-        from collective.cart.core.interfaces import IShoppingSite
-        from collective.cart.core.interfaces import IShoppingSiteRoot
-        from plone.dexterity.utils import createContentInContainer
-        from zope.interface import alsoProvides
-        from zope.lifecycleevent import modified
-
         folder = self.create_folder()
         alsoProvides(folder, IShoppingSiteRoot)
-
-        IShoppingSite(folder).update_next_cart_id()
+        adapter = IShoppingSite(folder)
+        adapter.update_next_cart_id()
         self.assertFalse(ICartContainerAdapter.called)
 
-        container = createContentInContainer(
-            folder, 'collective.cart.core.CartContainer', id='container', checkConstraints=False)
-        modified(container)
-
-        IShoppingSite(folder).update_next_cart_id()
+        self.create_cart_container(folder)
+        adapter.update_next_cart_id()
         self.assertTrue(ICartContainerAdapter.called)
+
+    def test_create_cart(self):
+        folder = self.create_folder()
+        alsoProvides(folder, IShoppingSiteRoot)
+        self.create_cart_container(folder)